# backend-go/docs/swagger.yaml
basePath: /
definitions:
  dto.ClaimAttachmentListResponse:
    properties:
      attachments:
        items:
          $ref: '#/definitions/entity.ClaimAttachment'
        type: array
      total:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserDTO'
    type: object
  dto.RefreshTokenResponse:
    properties:
      token:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
    type: object
...
  dto.UserDTO:
    properties:
      email:
        maxLength: 100
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      office_id:
        type: string
      role:
        maxLength: 20
        type: string
    required:
    - email
    - id
    - name
    - office_id
    - role
    type: object
  dto.ValidateTokenResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserDTO'
      valid:
        type: boolean
    type: object
...
  entity.ClaimAttachment:
    properties:
      claim_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ev-warranty.com
    name: API Support
  description: API for EV Warranty Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EV Warranty API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user by invalidating refresh token
      produces:
      - application/json
      responses:
        "204":
          description: Logout successful
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User logout
      tags:
      - auth
  /auth/token:
    get:
      consumes:
      - application/json
      description: Validate the provided JWT access token
      produces:
      - application/json
      responses:
        "200":
          description: Token is valid
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ValidateTokenResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Validate access token
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: New access token
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RefreshTokenResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
...
  /claims/{id}/attachments:
    get:
      consumes:
      - application/json
      description: Retrieve all attachments for a specific claim
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim attachments retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ClaimAttachmentListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get claim attachments by claim ID
      tags:
      - claim-attachments
    post:
      consumes:
      - multipart/form-data
      description: Upload files as attachments to a claim (SC Technician only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Files to upload
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Claim attachments uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.ClaimAttachment'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Upload claim attachments
      tags:
      - claim-attachments
  /claims/{id}/attachments/{attachmentID}:
    delete:
      consumes:
      - application/json
      description: Remove an attachment from a claim (SC Technician only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachmentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim attachment deleted successfully
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim attachment not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a claim attachment
      tags:
      - claim-attachments
    get:
      consumes:
      - application/json
      description: Retrieve a specific claim attachment by its ID
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachmentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim attachment retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.ClaimAttachment'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim attachment not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get claim attachment by ID
      tags:
      - claim-attachments
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"